+++ src/frontend/js/utils/test/factories/richie.ts
@@ -2,7 +2,14 @@ import { faker } from '@faker-js/faker';
 import { User } from 'types/User';
 import { APIBackend } from 'types/api';
 import { CommonDataProps } from 'types/commonDataProps';
-import { CourseRun, CourseRunDisplayMode, CourseState, CourseStateTextEnum, Priority } from 'types';
+import {
+  CourseRun,
+  CourseRunDisplayMode,
+  CourseState,
+  CourseStateTextEnum,
+  PacedCourse,
+  Priority,
+} from 'types';
 import { Course } from 'types/Course';
 import { FactoryHelper } from 'utils/test/factories/helper';
 import { factory } from './factories';
@@ -51,6 +58,14 @@ export const CourseRunFactory = factory<CourseRun>(() => {
     dashboard_link: null,
     title: faker.lorem.sentence(3),
     display_mode: CourseRunDisplayMode.DETAILED,
+  };
+});
+
+export const PacedCourseFactory = factory((): PacedCourse => {
+  return {
+    id: faker.string.uuid(),
+    code: faker.string.alphanumeric(5),
+    is_self_paced: false,
   };
 });